Rearrange pypaperless types
