+++ pypaperless/models/mixins/data_fields.py
@@ -1,13 +1,31 @@
 """MatchingFieldsMixin for PyPaperless models."""
 
+from enum import Enum
+
 from pydantic import BaseModel
 
-from pypaperless.models.common import MatchingAlgorithmType
+
+class MatchingAlgorithm(Enum):
+    """Represent a subtype of `Correspondent`, `DocumentType`, `StoragePath` and `Tag`."""
+
+    NONE = 0
+    ANY = 1
+    ALL = 2
+    LITERAL = 3
+    REGEX = 4
+    FUZZY = 5
+    AUTO = 6
+    UNKNOWN = -1
+
+    @classmethod
+    def _missing_(cls: type, *_: object) -> "MatchingAlgorithm":
+        """Set default member on unknown value."""
+        return MatchingAlgorithm.UNKNOWN
 
 
 class MatchingFieldsMixin(BaseModel):
     """Provide shared matching fields for PyPaperless models."""
 
     match: str | None = None
-    matching_algorithm: MatchingAlgorithmType | None = None
+    matching_algorithm: MatchingAlgorithm | None = None
     is_insensitive: bool | None = None
