+++ pypaperless/models/status.py
@@ -1,16 +1,72 @@
 """Provide `Status` related models."""
 
+import datetime
+from enum import Enum
 from typing import ClassVar, cast
 
+from pydantic import BaseModel, Field
+
 from pypaperless.const import API_PATH
-from pypaperless.models.common import (
-    StatusDatabaseType,
-    StatusStorageType,
-    StatusTasksType,
-    StatusType,
-)
 
 from .base import PaperlessModel
+
+
+class StatusType(Enum):
+    """Represent a subtype of `Status`."""
+
+    OK = "OK"
+    ERROR = "ERROR"
+    WARNING = "WARNING"
+    UNKNOWN = "UNKNOWN"
+
+    @classmethod
+    def _missing_(cls: type, *_: object) -> "StatusType":
+        """Set default member on unknown value."""
+        return StatusType.UNKNOWN
+
+
+class StatusDatabaseMigration(BaseModel):
+    """Represent a subtype of `StatusDatabase`."""
+
+    latest_migration: str | None = None
+    unapplied_migrations: list[str] = Field(default_factory=list)
+
+
+class StatusDatabase(BaseModel):
+    """Represent a subtype of `Status`."""
+
+    type: str | None = None
+    url: str | None = None
+    status: StatusType | None = None
+    error: str | None = None
+    migration_status: StatusDatabaseMigration | None = None
+
+
+class StatusStorage(BaseModel):
+    """Represent a subtype of `Status`."""
+
+    total: int | None = None
+    available: int | None = None
+
+
+class StatusTasks(BaseModel):
+    """Represent a subtype of `Status`."""
+
+    redis_url: str | None = None
+    redis_status: StatusType | None = None
+    redis_error: str | None = None
+    celery_status: StatusType | None = None
+    celery_url: str | None = None
+    celery_error: str | None = None
+    index_status: StatusType | None = None
+    index_last_modified: datetime.datetime | None = None
+    index_error: str | None = None
+    classifier_status: StatusType | None = None
+    classifier_last_trained: datetime.datetime | None = None
+    classifier_error: str | None = None
+    sanity_check_status: StatusType | None = None
+    sanity_check_last_run: datetime.datetime | None = None
+    sanity_check_error: str | None = None
 
 
 class Status(PaperlessModel):
@@ -21,9 +77,9 @@ class Status(PaperlessModel):
     pngx_version: str | None = None
     server_os: str | None = None
     install_type: str | None = None
-    storage: StatusStorageType | None = None
-    database: StatusDatabaseType | None = None
-    tasks: StatusTasksType | None = None
+    storage: StatusStorage | None = None
+    database: StatusDatabase | None = None
+    tasks: StatusTasks | None = None
 
     @property
     def has_errors(self) -> bool:
